Use double precision for clip planes to overcome JSON precision issues
The float-array to JSON conversion in staticjson uses double to float casting,
thus introducing unneccesary precision issues. Using double directly is the
best solution right now.

+++ brayns/common/types.h
@@ -332,7 +332,7 @@ enum class CameraMode
  * in absolute value of the coordinate system. Values are stored
  * in a Vector4, with the following order: nx, ny, nz and d
  */
-using ClipPlane = std::array<float, 4>;
+using ClipPlane = std::array<double, 4>;
 using ClipPlanes = std::vector<ClipPlane>;
 
 struct RenderInput
